refactor: 컴포넌트와 동일한 위치로 CSS 파일 통합 및 CSS Module 적용
