+++ src/components/Header/GNBBar/GNBBar.tsx
@@ -1,18 +1,20 @@
-import MenuLink from "@/components/common/MenuLink.tsx";
+import MenuLink from "@/components/common/MenuLink/MenuLink.tsx";
 
-import "@/components/Header/GNBBar/GNBBar.css";
+import style from "@/components/Header/GNBBar/GNBBar.module.css";
 
 interface GNBProps {
   onHover: React.MouseEventHandler;
 }
 
+const { s_nav, s_gnb, s_lnb } = style;
+
 const GNBBar = ({ onHover }: GNBProps) => {
   return (
-    <nav className="nav" onMouseEnter={onHover} onMouseLeave={onHover}>
-      <ul className="gnb">
+    <nav className={s_nav} onMouseEnter={onHover} onMouseLeave={onHover}>
+      <ul className={s_gnb}>
         <li>
           <a href="#">메뉴소개</a>
-          <ul className="lnb">
+          <ul className={s_lnb}>
             <MenuLink href="#">샌드위치</MenuLink>
             <MenuLink href="#">랩ㆍ기타</MenuLink>
             <MenuLink href="#">샐러드</MenuLink>
@@ -24,7 +26,7 @@ const GNBBar = ({ onHover }: GNBProps) => {
 
         <li>
           <a href="#">이용방법</a>
-          <ul className="lnb">
+          <ul className={s_lnb}>
             <MenuLink href="#">써브웨이 이용방법</MenuLink>
             <MenuLink href="#">단체메뉴 이용방법</MenuLink>
             <MenuLink href="#">신선한 재료 소개</MenuLink>
@@ -34,7 +36,7 @@ const GNBBar = ({ onHover }: GNBProps) => {
 
         <li>
           <a href="#">새소식</a>
-          <ul className="lnb">
+          <ul className={s_lnb}>
             <MenuLink href="#">이벤트ㆍ프로모션</MenuLink>
             <MenuLink href="#">뉴스ㆍ공지사항</MenuLink>
             <MenuLink href="#">광고영상</MenuLink>
@@ -43,7 +45,7 @@ const GNBBar = ({ onHover }: GNBProps) => {
 
         <li>
           <a href="#">쎄브웨이</a>
-          <ul className="lnb">
+          <ul className={s_lnb}>
             <MenuLink href="#">써브웨이 역사</MenuLink>
             <MenuLink href="#">써브웨이 약속</MenuLink>
             <MenuLink href="#">샌드위치 아티스트 지원</MenuLink>
@@ -53,7 +55,7 @@ const GNBBar = ({ onHover }: GNBProps) => {
 
         <li>
           <a href="#">가맹점</a>
-          <ul className="lnb">
+          <ul className={s_lnb}>
             <MenuLink href="#">써브웨이 프랜차이즈</MenuLink>
             <MenuLink href="#">가맹관련 FAQ</MenuLink>
             <MenuLink href="#">가맹신청ㆍ문의</MenuLink>
@@ -64,7 +66,7 @@ const GNBBar = ({ onHover }: GNBProps) => {
 
         <li>
           <a href="#">온라인 주문</a>
-          <ul className="lnb">
+          <ul className={s_lnb}>
             <MenuLink href="#">FAST-SUB</MenuLink>
             <MenuLink href="#">HOME-SUB</MenuLink>
             <MenuLink href="#">단체주문</MenuLink>
