+++ src/components/Header/Logo/logo.module.css
@@ -1,4 +1,4 @@
-.logo {
+.s_logo {
   width: 100%;
   padding: 30px 0;
   text-align: center;
