+++ src/components/Header/UtilMenu/UtilMenu.tsx
@@ -1,19 +1,21 @@
-import MenuLink from "@/components/common/MenuLink.tsx";
+import MenuLink from "@/components/common/MenuLink/MenuLink.tsx";
 
 import globalIcon from "@/assets/global_subway_icon.png";
 
-import "@/components/Header/UtilMenu/utilMenu.css";
+import style from "@/components/Header/UtilMenu/utilMenu.module.css";
+
+const { s_utilMenu, s_sign, s_globalSubwayIcon } = style;
 
 const UtilMenu = () => {
   return (
-    <ul className="util-menu">
-      <MenuLink href="#" className="sign">
+    <ul className={s_utilMenu}>
+      <MenuLink href="#" className={s_sign}>
         로그인
       </MenuLink>
-      <MenuLink href="#" className="sign">
+      <MenuLink href="#" className={s_sign}>
         회원가입
       </MenuLink>
-      <MenuLink href="#" className="global-subway-icon">
+      <MenuLink href="#" className={s_globalSubwayIcon}>
         <img src={globalIcon} alt="Global Subway" />
       </MenuLink>
     </ul>
