+++ src/main.tsx
@@ -3,8 +3,8 @@ import ReactDOM from "react-dom/client";
 
 import App from "@/App.tsx";
 
-import "@/styles/fonts.css";
-import "@/styles/reset.css";
+import "@/styles/_fonts.css";
+import "@/styles/_reset.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <StrictMode>
