+++ src/styles/_reset.css
@@ -1,8 +1,3 @@
-/* http://meyerweb.com/eric/tools/css/reset/ 
-   v2.0 | 20110126
-   License: none (public domain)
-*/
-
 html,
 body,
 div,
@@ -134,4 +129,9 @@ q::after {
 table {
   border-spacing: 0;
   border-collapse: collapse;
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
 }
